+++ src/components/Form/Success.jsx
@@ -1,9 +1,33 @@
-import {useStep, useCreate} from '../../hooks/useForm';
+import {useEffect} from 'react';
 
-const Success = () => {
-	const step = useStep(state => state.step);
-	const prevStep = useStep(state => state.prevStep);
-	return <div>Successfully created Event</div>;
+import {useCreate} from '../../hooks/useForm';
+
+const Success = ({error}) => {
+	const setNewEvent = useCreate(state => state.setNewEvent);
+
+	useEffect(() => {
+		setNewEvent({
+			pictures: {
+				eventPicture: '',
+			},
+			title: '',
+			private: false,
+			street: '',
+			city: '',
+			zip: '',
+			date: '',
+			time: '',
+			description: '',
+		});
+	}, []);
+
+	return (
+		<>
+			{!error && <div div>Successfully created Event</div>}
+
+			{error && <div>An error occured: Please try again later {error}</div>}
+		</>
+	);
 };
 
 export default Success;
