+++ src/hooks/useForm.js
@@ -7,7 +7,7 @@ export const useStep = create(
 			step: 1,
 			nextStep: () => set(state => ({step: state.step + 1})),
 			prevStep: () => set(state => ({step: state.step - 1})),
-			resetStep: () => set(state => ({step: 1})),
+			resetStep: () => set(() => ({step: 1})),
 		}),
 		{
 			name: 'step',
